👌🚚 [RUM] move manual error collection to domain/rumEventsCollection
