+++ packages/rum/src/boot/rum.spec.ts
@@ -601,75 +601,3 @@ describe('rum view url', () => {
     expect(getRumMessage(server, 0).view.url).toEqual('http://foo.com/')
   })
 })
-
-describe('rum manual error collection', () => {
-  let setupBuilder: TestSetupBuilder
-
-  beforeEach(() => {
-    setupBuilder = setup().withRum()
-  })
-
-  afterEach(() => {
-    setupBuilder.cleanup()
-  })
-
-  it('notifies a raw rum error event', () => {
-    const { lifeCycle, rawRumEvents } = setupBuilder.build()
-    lifeCycle.notify(LifeCycleEventType.MANUAL_ERROR_COLLECTED, {
-      error: {
-        error: new Error('foo'),
-        source: ErrorSource.CUSTOM,
-        startTime: 12,
-      },
-    })
-
-    expect(rawRumEvents.length).toBe(1)
-    expect(rawRumEvents[0]).toEqual({
-      customerContext: undefined,
-      rawRumEvent: {
-        date: jasmine.any(Number),
-        error: {
-          kind: 'Error',
-          origin: ErrorSource.CUSTOM,
-          stack: jasmine.stringMatching('Error: foo'),
-        },
-        evt: {
-          category: RumEventCategory.ERROR,
-        },
-        message: 'foo',
-      },
-      savedGlobalContext: undefined,
-      startTime: 12,
-    })
-  })
-
-  it('should save the specified customer context', () => {
-    const { lifeCycle, rawRumEvents } = setupBuilder.build()
-    lifeCycle.notify(LifeCycleEventType.MANUAL_ERROR_COLLECTED, {
-      error: {
-        context: { foo: 'bar' },
-        error: new Error('foo'),
-        source: ErrorSource.CUSTOM,
-        startTime: 12,
-      },
-    })
-    expect(rawRumEvents[0].customerContext).toEqual({
-      foo: 'bar',
-    })
-  })
-
-  it('should save the global context', () => {
-    const { lifeCycle, rawRumEvents } = setupBuilder.build()
-    lifeCycle.notify(LifeCycleEventType.MANUAL_ERROR_COLLECTED, {
-      context: { foo: 'bar' },
-      error: {
-        error: new Error('foo'),
-        source: ErrorSource.CUSTOM,
-        startTime: 12,
-      },
-    })
-    expect(rawRumEvents[0].savedGlobalContext).toEqual({
-      foo: 'bar',
-    })
-  })
-})
