+++ packages/rum/src/domain/lifeCycle.ts
@@ -1,7 +1,8 @@
 import { Context, ErrorMessage } from '@datadog/browser-core'
 import { RumPerformanceEntry } from '../browser/performanceCollection'
-import { ManuallyAddedError, RawRumEvent, RumEvent } from '../types'
+import { RawRumEvent, RumEvent } from '../types'
 import { RawRumEventV2, RumEventV2 } from '../typesV2'
+import { ManuallyAddedError } from './rumEventsCollection/errorCollection'
 import { RequestCompleteEvent, RequestStartEvent } from './rumEventsCollection/requestCollection'
 import { AutoActionCreatedEvent, AutoUserAction, CustomUserAction } from './rumEventsCollection/userActionCollection'
 import { View, ViewCreatedEvent } from './rumEventsCollection/viewCollection'
